#128 - Change int -> long for IDs

+++ src/ZCrew.StateCraft/StateMachines/StateMachine.cs
@@ -25,7 +25,7 @@ internal sealed partial class StateMachine<TState, TTransition> : StateMachineBa
     private readonly IReadOnlyList<ITrigger> triggers;
     private InternalState internalState;
 
-    private readonly int id;
+    private readonly long id;
     private readonly AsyncLock stateMachineLock = new();
     private Task? actionTask;
     private CancellationTokenSource? actionCancellationTokenSource;
@@ -204,7 +204,7 @@ internal sealed partial class StateMachine<TState, TTransition> : StateMachineBa
             var task = this.actionTask;
             this.actionTask = null;
 
-            var stateMachineIds = AsynchronousStateMachineIds.Value ?? ImmutableHashSet<int>.Empty;
+            var stateMachineIds = AsynchronousStateMachineIds.Value ?? ImmutableHashSet<long>.Empty;
 
             // Since the asynchronous action is interacting with the state machine we can compensate for a user-issue by
             // avoiding cancellation until that action has been completed
@@ -273,7 +273,7 @@ internal sealed partial class StateMachine<TState, TTransition> : StateMachineBa
             var actionCts = new CancellationTokenSource();
 
             // Add this ID to the call chain to avoid deadlocks on ExitState
-            var stateMachineIds = AsynchronousStateMachineIds.Value ?? ImmutableHashSet<int>.Empty;
+            var stateMachineIds = AsynchronousStateMachineIds.Value ?? ImmutableHashSet<long>.Empty;
             AsynchronousStateMachineIds.Value = stateMachineIds.Add(this.id);
 
             // Start the execution of the action and allow the state machine to be transitioned or deactivated.
